Use Multiple Resolvers by Default
Currently the only way to successfully pass `-r` argument is by editing the module file directly. If provided in cmd, it fails trying to create an output file named `-r`. Setting it to use multiple resolvers by default is a quick fix to the parsing issue.

Consider setting a best practice for additional arguments passed to binaries, maybe adding in the module makes more sense.